🚚️ api: move serializer to src/enrolment context
